tests: split test in separate file
