+++ crates/meilisearch/src/routes/indexes/search.rs
@@ -198,7 +198,7 @@ impl TryFrom<SearchQueryGet> for SearchQuery {
 // TODO: TAMO: split on :asc, and :desc, instead of doing some weird things
 
 /// Transform the sort query parameter into something that matches the post expected format.
-fn fix_sort_query_parameters(sort_query: &str) -> Vec<String> {
+pub fn fix_sort_query_parameters(sort_query: &str) -> Vec<String> {
     let mut sort_parameters = Vec::new();
     let mut merge = false;
     for current_sort in sort_query.trim_matches('"').split(',').map(|s| s.trim()) {
@@ -354,32 +354,5 @@ pub fn search_kind(
         ),
 
         (_, None, Some(_)) => Err(MeilisearchHttpError::MissingSearchHybrid.into()),
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn test_fix_sort_query_parameters() {
-        let sort = fix_sort_query_parameters("_geoPoint(12, 13):asc");
-        assert_eq!(sort, vec!["_geoPoint(12,13):asc".to_string()]);
-        let sort = fix_sort_query_parameters("doggo:asc,_geoPoint(12.45,13.56):desc");
-        assert_eq!(sort, vec!["doggo:asc".to_string(), "_geoPoint(12.45,13.56):desc".to_string(),]);
-        let sort = fix_sort_query_parameters(
-            "doggo:asc , _geoPoint(12.45, 13.56, 2590352):desc , catto:desc",
-        );
-        assert_eq!(
-            sort,
-            vec![
-                "doggo:asc".to_string(),
-                "_geoPoint(12.45,13.56,2590352):desc".to_string(),
-                "catto:desc".to_string(),
-            ]
-        );
-        let sort = fix_sort_query_parameters("doggo:asc , _geoPoint(1, 2), catto:desc");
-        // This is ugly but eh, I don't want to write a full parser just for this unused route
-        assert_eq!(sort, vec!["doggo:asc".to_string(), "_geoPoint(1,2),catto:desc".to_string(),]);
     }
 }
